+++ spock-specs/src/test/groovy/org/spockframework/smoke/ast/AstSpec.groovy
@@ -387,158 +387,4 @@ class TestSpec extends Specification {
     then:
     textSnapshotter.assertThat(result.source).matchesSnapshot()
   }
-
-  def "GDK method that looks like built-in method as implicit condition"() {
-    when:
-    def result = compiler.transpileFeatureBody('''
-expect:
-null.with {
-  false
-}
-''')
-
-    then:
-    snapshotter.assertThat(result.source).matchesSnapshot()
-  }
-
-  def "GDK method that looks like built-in method as explicit condition"() {
-    when:
-    def result = compiler.transpileFeatureBody('''
-expect:
-assert null.with {
-  false
-}
-''')
-
-    then:
-    snapshotter.assertThat(result.source).matchesSnapshot()
-  }
-
-  def "condition method #conditionMethod within condition method #conditionMethod"() {
-    when:
-    def result = compiler.transpileFeatureBody("""
-expect:
-$conditionMethod(['']) {
-  $conditionMethod(['']) {
-    false
-  }
-}
-""")
-
-    then:
-    snapshotter.assertThat(result.source).matchesSnapshot()
-
-    where:
-    conditionMethod << [
-      'with',
-      'verifyAll',
-      'verifyEach'
-    ]
-  }
-
-  def "condition method #conditionMethod within condition method #conditionMethod with exception"() {
-    when:
-    def result = compiler.transpileFeatureBody("""
-expect:
-$conditionMethod(['']) {
-  $conditionMethod(['']) {
-    true
-    throw new Exception('foo')
-  }
-}
-""")
-
-    then:
-    snapshotter.assertThat(result.source).matchesSnapshot()
-
-    where:
-    conditionMethod << [
-      'with',
-      'verifyAll',
-      'verifyEach'
-    ]
-  }
-
-  def "condition method #conditionMethod within condition method #conditionMethod with only exception"() {
-    when:
-    def result = compiler.transpileFeatureBody("""
-expect:
-$conditionMethod(['']) {
-  $conditionMethod(['']) {
-    throw new Exception('foo')
-  }
-}
-""")
-
-    then:
-    snapshotter.assertThat(result.source).matchesSnapshot()
-
-    where:
-    conditionMethod << [
-      'with',
-      'verifyAll',
-      'verifyEach'
-    ]
-  }
-
-  def "condition method #conditionMethod"() {
-    when:
-    def result = compiler.transpileFeatureBody("""
-expect:
-$conditionMethod(['']) {
-  false
-}
-""")
-
-    then:
-    snapshotter.assertThat(result.source).matchesSnapshot()
-
-    where:
-    conditionMethod << [
-      'with',
-      'verifyAll',
-      'verifyEach'
-    ]
-  }
-
-  def "condition method #conditionMethod with exception"() {
-    when:
-    def result = compiler.transpileFeatureBody("""
-expect:
-$conditionMethod(['']) {
-  true
-  throw new Exception('foo')
-}
-""")
-
-    then:
-    snapshotter.assertThat(result.source).matchesSnapshot()
-
-    where:
-    conditionMethod << [
-      'with',
-      'verifyAll',
-      'verifyEach'
-    ]
-  }
-
-  def "condition method #conditionMethod with only exception"() {
-    when:
-    def result = compiler.transpileFeatureBody("""
-expect:
-$conditionMethod(['']) {
-  throw new Exception('foo')
-}
-""")
-
-    then:
-    snapshotter.assertThat(result.source).matchesSnapshot()
-
-    where:
-    conditionMethod << [
-      'with',
-      'verifyAll',
-      'verifyEach'
-    ]
-  }
 }
